lumin-backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-allauth==0.57.0

# Database
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.4
redis==5.0.1

# Configuration
django-environ==0.11.2

# File Storage
boto3==1.34.10

# External Services
twilio==9.0.0
sendgrid==6.10.0

# Server
gunicorn==21.2.0
whitenoise==6.6.0

# CORS
django-cors-headers==4.3.1

# Security & Rate Limiting
django-ratelimit==4.1.0
cryptography==41.0.7

# Utilities
django-filter==23.5
Pillow==10.2.0
python-dateutil==2.8.2

# Development
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
